add deploy configs

+++ src/main/kotlin/com/fantasy/football/service/TelegramService.kt
@@ -10,7 +10,7 @@ import mu.KotlinLogging
 class TelegramService(private val bot: Bot, private val telegramConfig: TelegramConfig) {
 
     fun sendMessage(string: String) {
-        val chatId = ChatId.fromId(telegramConfig.chatId)
+        val chatId = ChatId.fromId(telegramConfig.chatId.toLong())
         val newString = escapeInvalidChars(string)
 
         val messageResponse =
